requirements.txt
streamlit==1.24.0
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
openai==0.28.1
altair==4.2.2
python-dotenv==1.0.0 